+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Implemented enhancements
 - TT-1619 - Sending New Registration Letters new env FEATURE_REGISTRATION_LETTERS_ENABLED
 - no ticket - Increase flake8 Char limit to 120
+- TT-851 - Companies house status check admin upgrade
 - TT-1613 Reduce stannp to flag errors from <50 to <10 
 
 ## [2019.07.15](https://github.com/uktrade/directory-api/releases/tag/2019.07.15)